新增以token请求头形式解析的方法getUserDetailedInfo

+++ src/main/java/com/example/musicbox/common/UserInfo.java
@@ -4,13 +4,13 @@ package com.example.musicbox.common;
  * 在解析token后用于存储用户id的ThreadLocal
  */
 public class UserInfo {
-    private static final ThreadLocal<Integer> userThreadLocal = new ThreadLocal<>();
+    private static final ThreadLocal<Long> userThreadLocal = new ThreadLocal<>();
 
-    public static void set(Integer userId){
+    public static void set(long userId){
         userThreadLocal.set(userId);
     }
 
-    public static Integer get(){
+    public static long get(){
         return userThreadLocal.get();
     }
